Moved Rosenbrock support classes to Native folder
