+++ GridKit/Solver/Dynamic/CMakeLists.txt
@@ -57,24 +57,6 @@ if(GRIDKIT_ENABLE_RESOLVE)
       GridKit::utilities_logger
       PUBLIC
       GridKit::sparse_matrix)
-
-  gridkit_add_library(
-    inf_norm
-    SOURCES InfNorm.cpp
-    HEADERS InfNorm.hpp
-    LINK_LIBRARIES
-      PRIVATE
-      ReSolve::ReSolve
-      PUBLIC
-      GridKit::dense_vector)
 endif()
 
-gridkit_add_library(
-  adaptive_step
-  SOURCES AdaptiveStep.cpp
-  HEADERS AdaptiveStep.hpp)
-
-gridkit_add_library(
-  fixed_step
-  SOURCES FixedStep.cpp
-  HEADERS FixedStep.hpp)
+add_subdirectory(Native)
